Rename files of nested types.
